Move globsToMatcher from jest-core to jest-util
I'd like to start using this in more places to improve performance.
Moving it to jest-util seems like a better spot. Now that it is a
standalone module, I decided to write some unit tests for this
function. In doing so, I uncovered a small difference between the
behavior of this function and micromatch when overlapping glob
patterns are used which I also fixed.
