+++ packages/jest-util/package.json
@@ -14,11 +14,13 @@
     "chalk": "^4.0.0",
     "graceful-fs": "^4.2.4",
     "is-ci": "^2.0.0",
-    "make-dir": "^3.0.0"
+    "make-dir": "^3.0.0",
+    "micromatch": "^4.0.2"
   },
   "devDependencies": {
     "@types/graceful-fs": "^4.1.2",
     "@types/is-ci": "^2.0.0",
+    "@types/micromatch": "^4.0.0",
     "@types/node": "*"
   },
   "engines": {
